Added check of FOVTYPE presence to check_flux().
Now complains if FOVTYPE (but not FOV) is present and CALSTAT is
'U'. Doesn't require FOVTYPE to be present if CALSTAT is 'C', which
seems to be consistent with the draft standard.

Changed oitable-demo so that ARRNAME, FOV, FOVTYPE and the STA_INDEX
column are optional, to allow generation of compliant OIFITS files.